fix StrongID yaml support
